YHRM/requirements.txt
diff-match-patch==20121119
Django==1.8.17
django-debugtools==1.7.2
django-import-export==0.5.1
django-report-builder==3.4.4
django-simple-history==1.8.2
django-suit==0.2.24
djangorestframework==3.5.4
et-xmlfile==1.0.1
jdcal==1.3
MySQL-python==1.2.3
openpyxl==2.3.5
python-dateutil==2.6.0
six==1.10.0
tablib==0.11.4
django-admin-bootstrapped
django-filter
